Change namespace of removeBoldTagWrapper filter.
